Added new command to quill, and minimal docs.
